requirements.txt
requests
aiogram
python-dotenv
psycopg2-binary